requirements.txt
numpy==1.18.5
uvicorn==0.11.5
fastapi==0.56.0
tensorflow==2.1.0
transformers==2.11.0
PyYAML==5.3.1
scikit-learn==0.21.3

azure-storage-blob==12.3.2
azureml==0.2.7
azureml-train==1.8.0